requirements.txt
python-binance
numpy
backtrader
flask
pandas~=1.2.2
plotly~=4.14.3
python-crontab~=2.5.1
python-dateutil~=2.8.1
fastapi
uvicorn
jinja2